# Holds wasm.Code as values on wasm.Module