--- a/imports/wasi_snapshot_preview1/proc.go
+++ b/imports/wasi_snapshot_preview1/proc.go
@@ -23,7 +23,7 @@ var procExit = &wasm.HostFunc{
 	Name:        ProcExitName,
 	ParamTypes:  []api.ValueType{i32},
 	ParamNames:  []string{"rval"},
-	Code: &wasm.Code{
+	Code: wasm.Code{
 		GoFunc: api.GoModuleFunc(procExitFn),
 	},
 }