--- a/internal/integration_test/fuzz/wazerolib/nodiff.go
+++ b/internal/integration_test/fuzz/wazerolib/nodiff.go
@@ -179,7 +179,7 @@ func ensureDummyImports(r wazero.Runtime, origin *wasm.Module, requireNoError fu
 					}
 				}
 				body.WriteByte(wasm.OpcodeEnd)
-				m.CodeSection = append(m.CodeSection, &wasm.Code{Body: body.Bytes()})
+				m.CodeSection = append(m.CodeSection, wasm.Code{Body: body.Bytes()})
 			case wasm.ExternTypeGlobal:
 				index = uint32(len(m.GlobalSection))
 				var data []byte