--- a/internal/wasm/binary/code.go
+++ b/internal/wasm/binary/code.go
@@ -10,10 +10,10 @@ import (
 	"github.com/tetratelabs/wazero/internal/wasm"
 )
 
-func decodeCode(r *bytes.Reader, codeSectionStart uint64) (*wasm.Code, error) {
+func decodeCode(r *bytes.Reader, codeSectionStart uint64, ret *wasm.Code) (err error) {
 	ss, _, err := leb128.DecodeUint32(r)
 	if err != nil {
-		return nil, fmt.Errorf("get the size of code: %w", err)
+		return fmt.Errorf("get the size of code: %w", err)
 	}
 	remaining := int64(ss)
 
@@ -21,9 +21,9 @@ func decodeCode(r *bytes.Reader, codeSectionStart uint64) (*wasm.Code, error) {
 	ls, bytesRead, err := leb128.DecodeUint32(r)
 	remaining -= int64(bytesRead)
 	if err != nil {
-		return nil, fmt.Errorf("get the size locals: %v", err)
+		return fmt.Errorf("get the size locals: %v", err)
 	} else if remaining < 0 {
-		return nil, io.EOF
+		return io.EOF
 	}
 
 	var nums []uint64
@@ -34,9 +34,9 @@ func decodeCode(r *bytes.Reader, codeSectionStart uint64) (*wasm.Code, error) {
 		n, bytesRead, err = leb128.DecodeUint32(r)
 		remaining -= int64(bytesRead) + 1 // +1 for the subsequent ReadByte
 		if err != nil {
-			return nil, fmt.Errorf("read n of locals: %v", err)
+			return fmt.Errorf("read n of locals: %v", err)
 		} else if remaining < 0 {
-			return nil, io.EOF
+			return io.EOF
 		}
 
 		sum += uint64(n)
@@ -44,19 +44,19 @@ func decodeCode(r *bytes.Reader, codeSectionStart uint64) (*wasm.Code, error) {
 
 		b, err := r.ReadByte()
 		if err != nil {
-			return nil, fmt.Errorf("read type of local: %v", err)
+			return fmt.Errorf("read type of local: %v", err)
 		}
 		switch vt := b; vt {
 		case wasm.ValueTypeI32, wasm.ValueTypeF32, wasm.ValueTypeI64, wasm.ValueTypeF64,
 			wasm.ValueTypeFuncref, wasm.ValueTypeExternref, wasm.ValueTypeV128:
 			types = append(types, vt)
 		default:
-			return nil, fmt.Errorf("invalid local type: 0x%x", vt)
+			return fmt.Errorf("invalid local type: 0x%x", vt)
 		}
 	}
 
 	if sum > math.MaxUint32 {
-		return nil, fmt.Errorf("too many locals: %d", sum)
+		return fmt.Errorf("too many locals: %d", sum)
 	}
 
 	var localTypes []wasm.ValueType
@@ -70,12 +70,15 @@ func decodeCode(r *bytes.Reader, codeSectionStart uint64) (*wasm.Code, error) {
 	bodyOffsetInCodeSection := codeSectionStart - uint64(r.Len())
 	body := make([]byte, remaining)
 	if _, err = io.ReadFull(r, body); err != nil {
-		return nil, fmt.Errorf("read body: %w", err)
+		return fmt.Errorf("read body: %w", err)
 	}
 
 	if endIndex := len(body) - 1; endIndex < 0 || body[endIndex] != wasm.OpcodeEnd {
-		return nil, fmt.Errorf("expr not end with OpcodeEnd")
+		return fmt.Errorf("expr not end with OpcodeEnd")
 	}
 
-	return &wasm.Code{Body: body, LocalTypes: localTypes, BodyOffsetInCodeSection: bodyOffsetInCodeSection}, nil
+	ret.BodyOffsetInCodeSection = bodyOffsetInCodeSection
+	ret.LocalTypes = localTypes
+	ret.Body = body
+	return nil
 }