… internal/engine/interpreter/interpreter_test.go
… internal/integration_test/bench/hostfunc_bench_test.go
… internal/integration_test/engine/adhoc_test.go
… internal/testing/binaryencoding/encoder_test.go
… internal/wasm/counts_test.go
… internal/wasm/func_validation_test.go
… internal/wasm/function_definition_test.go
… internal/wasm/host_test.go
… internal/wasm/module_test.go
… internal/wasm/store_test.go
… internal/wasm/table_test.go
… internal/wazeroir/compiler_test.go
… runtime_test.go