```
$ benchstat old.txt new.txt
name                                    old time/op    new time/op    delta
Initialization/interpreter-32             35.9µs ± 1%    35.9µs ± 2%    ~     (p=0.528 n=25+26)
Initialization/interpreter-multiple-32    39.7µs ± 8%    39.6µs ± 9%    ~     (p=0.668 n=30+29)
Initialization/compiler-32                29.4µs ± 7%    28.8µs ± 6%  -2.06%  (p=0.014 n=28+30)
Initialization/compiler-multiple-32       25.3µs ±12%    25.7µs ± 9%    ~     (p=0.329 n=30+30)
Compilation/with_extern_cache-32           198µs ± 2%     201µs ± 2%  +1.14%  (p=0.000 n=28+30)
Compilation/without_extern_cache-32       5.97ms ± 1%    5.88ms ± 1%  -1.63%  (p=0.000 n=30+30)

name                                    old alloc/op   new alloc/op   delta
Initialization/interpreter-32              136kB ± 0%     136kB ± 0%  -0.00%  (p=0.008 n=19+30)
Initialization/interpreter-multiple-32     137kB ± 0%     137kB ± 0%    ~     (p=0.949 n=26+27)
Initialization/compiler-32                 137kB ± 0%     137kB ± 0%    ~     (p=0.313 n=30+30)
Initialization/compiler-multiple-32        142kB ± 0%     142kB ± 0%  -0.00%  (p=0.018 n=23+24)
Compilation/with_extern_cache-32          55.2kB ± 0%    54.6kB ± 0%  -1.00%  (p=0.000 n=30+30)
Compilation/without_extern_cache-32       1.99MB ± 0%    1.99MB ± 0%  -0.03%  (p=0.000 n=30+30)

name                                    old allocs/op  new allocs/op  delta
Initialization/interpreter-32               38.0 ± 0%      38.0 ± 0%    ~     (all equal)
Initialization/interpreter-multiple-32      57.0 ± 0%      57.0 ± 0%    ~     (all equal)
Initialization/compiler-32                  38.0 ± 0%      38.0 ± 0%    ~     (all equal)
Initialization/compiler-multiple-32         47.0 ± 0%      47.0 ± 0%    ~     (all equal)
Compilation/with_extern_cache-32           1.02k ± 0%     0.98k ± 0%  -3.96%  (p=0.000 n=28+30)
Compilation/without_extern_cache-32        32.7k ± 0%     32.6k ± 0%  -0.13%  (p=0.000 n=30+30)

```